REFACTOR: 컴포넌트 디렉토리 이동
- 출석 페이지 관련 컴포넌트들을 components 디렉토리의 attendance 디렉토리로 이동
